test(interpreter): store a value of the declared type past the alloca
The simplified test stored an `i64` constant under the type `i32`, so
the parser rejected it before the interpreter ran. The store is now an
`i64` store, which still runs past the 8-byte `alloca`.

+++ Test/Interpreter/LLVM/alloca_out_of_bounds.mlir
@@ -8,7 +8,7 @@
     %eight = "llvm.mlir.constant"() <{value = 8 : i64}> : () -> i64
     %p = "llvm.alloca"(%one) <{elem_type = i64}> : (i64) -> !llvm.ptr
     %p1 = "llvm.getelementptr"(%p, %eight) <{elem_type = i8, rawConstantIndices = array<i32: -2147483648>}> : (!llvm.ptr, i64) -> !llvm.ptr
-    "llvm.store"(%eight, %p1) : (i32, !llvm.ptr) -> ()
+    "llvm.store"(%eight, %p1) : (i64, !llvm.ptr) -> ()
     "func.return"() : () -> ()
   }) : () -> ()
 }) : () -> ()
